+++ src/hooks/utils/__tests__/useMouseAndTouchTracker.test.ts
@@ -1,6 +1,6 @@
 import {renderHook} from '@testing-library/react'
 import {useMouseAndTouchTracker} from '../useMouseAndTouchTracker'
-import {Environment} from '../../../downshift.types'
+import {Environment} from '../../../types'
 import {targetWithinDownshift} from '../../../utils'
 
 jest.mock('../../../utils', () => ({
